# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "Store API Collect&Verything",
        "version": "0.1"
    },
    "paths": {
...
        "/api/stores/{id}": {
            "get": {
                "tags": [
                    "Stores"
                ],
                "summary": "Get store information",
                "description": "Returns store data",
                "operationId": "retrieve",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "Store id",
                        "required": true
                    },
                    {
                        "name": "locale",
                        "in": "path",
                        "description": "Locale",
                        "required": false
                    }
                ],
                "responses": {
                    "200": {
                        "description": "successful operation"
                    },
                    "400": {
                        "description": "Bad request"
                    },
                    "404": {
                        "description": "Account not found."
                    },
                    "409": {
                        "description": "Conflict"
                    },
                    "500": {
                        "description": "Servor Error"
                    }
                }
            },
            "delete": {
                "tags": [
                    "Stores"
                ],
                "summary": "Delete a store",
                "description": "Soft delete a store",
                "operationId": "delete",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "Store id",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Account deleted"
                    },
                    "400": {
                        "description": "Bad request"
                    },
                    "404": {
                        "description": "Resource Not Found"
                    }
                }
            },
            "patch": {
                "tags": [
                    "Stores"
                ],
                "summary": "Patch a store",
                "description": "Update an store",
                "operationId": "update",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "Store id",
                        "required": true
                    },
                    {
                        "name": "account_id",
                        "in": "query",
                        "description": "Account Id",
                        "required": false
                    },
                    {
                        "name": "name",
                        "in": "query",
                        "description": "Store name",
                        "required": false
                    },
                    {
                        "name": "business_name",
                        "in": "query",
                        "description": "Store business Name",
                        "required": false
                    },
                    {
                        "name": "address_id",
                        "in": "query",
                        "description": "Address Id",
                        "required": false
                    },
                    {
                        "name": "phone",
                        "in": "query",
                        "description": "Store phone",
                        "required": false
                    },
                    {
                        "name": "email",
                        "in": "query",
                        "description": "Store email",
                        "required": false
                    },
                    {
                        "name": "type",
                        "in": "query",
                        "description": "Store type",
                        "required": false
                    },
                    {
                        "name": "openings",
                        "in": "query",
                        "description": "Store openings",
                        "required": false
                    },
                    {
                        "name": "primary_color",
                        "in": "query",
                        "description": "Store primary color",
                        "required": false
                    },
                    {
                        "name": "secondary_color",
                        "in": "query",
                        "description": "Store secondary color",
                        "required": false
                    },
                    {
                        "name": "logo",
                        "in": "query",
                        "description": "Store logo",
                        "required": false
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Account updated"
                    },
                    "400": {
                        "description": "Bad request"
                    },
                    "404": {
                        "description": "Resource Not Found"
                    }
                }
            }
        },
        "/api/stores": {
            "get": {
                "tags": [
                    "Stores"
                ],
                "summary": "Get all stores information",
                "description": "Returns store data",
                "operationId": "list",
                "parameters": [
                    {
                        "name": "locale",
                        "in": "path",
                        "description": "Locale",
                        "required": false
                    }
                ],
                "responses": {
                    "200": {
                        "description": "successful operation"
                    },
                    "400": {
                        "description": "Bad request"
                    },
                    "403": {
                        "description": "Forbidden"
                    },
                    "404": {
                        "description": "Resource Not Found"
                    },
                    "409": {
                        "description": "Conflict"
                    },
                    "500": {
                        "description": "Servor Error"
                    }
                }
            },
            "post": {
                "tags": [
                    "Stores"
                ],
                "summary": "Post a new store",
                "description": "Create a new store",
                "operationId": "create",
                "parameters": [
                    {
                        "name": "account_id",
                        "in": "query",
                        "description": "Account Id",
                        "required": true
                    },
                    {
                        "name": "name",
                        "in": "query",
                        "description": "Store name",
                        "required": true
                    },
                    {
                        "name": "business_name",
                        "in": "query",
                        "description": "Store business name",
                        "required": true
                    },
                    {
                        "name": "address_id",
                        "in": "query",
                        "description": "Address Id",
                        "required": true
                    },
                    {
                        "name": "phone",
                        "in": "query",
                        "description": "Store phone",
                        "required": true
                    },
                    {
                        "name": "email",
                        "in": "query",
                        "description": "Store email",
                        "required": true
                    },
                    {
                        "name": "type",
                        "in": "query",
                        "description": "Store type",
                        "required": true
                    },
                    {
                        "name": "openings",
                        "in": "query",
                        "description": "Store openings",
                        "required": true
                    },
                    {
                        "name": "primary_color",
                        "in": "query",
                        "description": "Store primary color",
                        "required": true
                    },
                    {
                        "name": "secondary_color",
                        "in": "query",
                        "description": "Store secondary color",
                        "required": true
                    },
                    {
                        "name": "logo",
                        "in": "query",
                        "description": "Store logo",
                        "required": true
                    },
                    {
                        "name": "locale",
                        "in": "query",
                        "description": "Locale",
                        "required": true
                    },
                    {
                        "name": "description",
                        "in": "query",
                        "description": "Description",
                        "required": true
                    }
                ],
                "responses": {
                    "201": {
                        "description": "Account created"
                    },
                    "400": {
                        "description": "Bad request"
                    },
                    "404": {
                        "description": "Resource Not Found"
                    }
                }
            }
        },
...
    },
    "components": {
        "securitySchemes": {
            "bearer_token": {
                "type": "apiKey",
                "description": "Enter token in format (Bearer <token>)",
                "name": "Authorization",
                "in": "header"
            }
        }
    },
    "security": [
        []
    ]
}